requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
pydantic
pydantic-settings
aiofiles
sqlalchemy
aiosqlite
loguru
httpx
groq
anthropic
gtts
ffmpeg-python
pillow
pymupdf
